Code refactoring Bumped up immunization library version
