+++ opensrp-chw-core/build.gradle
@@ -165,7 +165,7 @@ dependencies {
         exclude group: 'com.android.support', module: 'appcompat-v7'
     }
 
-    api('org.smartregister:opensrp-client-immunization:3.0.2-SNAPSHOT@aar') {
+    api('org.smartregister:opensrp-client-immunization:3.0.3-SNAPSHOT@aar') {
         transitive = true
         exclude group: 'org.smartregister', module: 'opensrp-client-core'
         exclude group: 'org.smartregister', module: 'opensrp-client-native-form'
