+++ opensrp-chw-core/src/main/java/org/smartregister/chw/core/dao/EventDao.java
@@ -109,28 +109,4 @@ public class EventDao extends AbstractDao {
             return false;
         }
     }
-
-    public static String getVaccineName(String formSubmissionID) {
-        try {
-            String sql = "SELECT name vaccineName FROM vaccines WHERE formSubmissionId='" + formSubmissionID + "';";
-
-            AbstractDao.DataMap<String> dataMap = cursor -> getCursorValue(cursor, "vaccineName");
-
-            return readSingleValue(sql, dataMap);
-        } catch (Exception e) {
-            return "";
-        }
-    }
-
-    public static String getVaccineDate(String formSubmissionID) {
-        try {
-            String sql = "SELECT date vaccineDate FROM vaccines WHERE formSubmissionId='" + formSubmissionID + "';";
-
-            AbstractDao.DataMap<String> dataMap = cursor -> getCursorValue(cursor, "vaccineDate");
-
-            return readSingleValue(sql, dataMap);
-        } catch (Exception e) {
-            return "";
-        }
-    }
 }
